requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pytz==2023.3
streamlit==1.52.1
pandas==2.3.3
plotly==6.5.0
